Feat/290 doughnut self contained tests (#768)
* Add self contained tests

* Add test and refactor accounts in tests

* Add tests for get_address and run_doughnut_common_validations

* Add more signed validation tests

+++ pallet/common/src/test_utils.rs
@@ -131,7 +131,7 @@ macro_rules! impl_frame_system_config {
 			type BlockLength = ();
 			type BaseCallFilter = frame_support::traits::Everything;
 			type RuntimeOrigin = RuntimeOrigin;
-			type Index = u64;
+			type Index = u32;
 			type BlockNumber = BlockNumber;
 			type RuntimeCall = RuntimeCall;
 			type Hash = H256;
